+++ packages/orchestrator/src/config.ts
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 */
-import composerUrl from '@micro-lc/composer/composer.production?url'
+import composerUri from '@micro-lc/composer/composer.production?url'
 import type {
   Application,
   Config,
@@ -60,7 +60,7 @@ export const defaultConfig: CompleteConfig = {
         integrationMode: 'parcel',
       },
     },
-    composerUri: composerUrl,
+    composerUri,
     defaultUrl: './',
   },
   shared: { properties: {} },